requirements.txt
flask
celery
redis
dash
plotly
flask_sqlalchemy
psycopg2-binary
python-dotenv